[tf.data] Add new custom transformation: `tf.contrib.data.scan()`.
`scan()` is similar to `Dataset.map()`, with the addition of a generic piece of
state that is accumulated across the elements of the input, and that may be
used in the computation of the output elements.

This change also updates `rejection_resample()` to use `scan()` rather than a
local `tf.ResourceVariable` for accumulating the number of times each class
has been encountered.

PiperOrigin-RevId: 171542274