[XLA:GPU] Use more threads per thread block.
Before this change, we supported two algorithms for choosing the number
of threads per block:

 * "optimize-for-latency" algorithm assumed that each thread would want
   the maximum number of registers it could have, and chose a block size
   small enough to accommodate this.
 * "optimize-for-throughput" algorithm packed as many threads into a
   block as possible.

In practice we always chose the optimize-for-latency algorithm.

This change removes the choice of algorithm and changes us to
unconditionally use a new one.  In our new algorithm, we choose the
smallest block size that still has the potential to allow the GPU to
reach maximum occupancy.

When each thread's register usage is small, we can pack many of these
blocks into one SM and hit maximum occupancy.  When the threads'
register usage is larger, we degrade gracefully (unlike with larger
block sizes, where the occupancy degredation is more quantized).

On our benchmarks, this is a moderate (0-10%) speedup on K40, and a
large (10-25%) speedup on P100.

PiperOrigin-RevId: 177879741