[XLA:GPU] Support atomic operations on small data types.
Implement atomic operations on less than 32 bit data types with atomicCAS(int).
This relies on an assumption that tensor buffers are 4 byte aligned and have a
size of 4N bytes in order not to access the buffers out of bound.

Enable the tests that requires atomic operations on small data types.

PiperOrigin-RevId: 178342413