Add `init_scope`, a scope for wrapping variable creation and initialization.
There is often a need to lift variable initialization ops out of control
flow contexts, graphs that are building functions, and gradient tapes.
Entering an `init_scope` is a mechanism for satisfying these desiderata. In
particular, entering an `init_scope` has three effects:

  (1) All control dependencies are cleared the moment the scope is entered;
      this is equivalent to entering the manager returned from
      `control_dependencies(None)`, which is how we exit control-flow contexts like
      `tf.while_loop` and `tf.cond`.
  (2) All operations that are created while the scope is active are lifted
      into the lowest context on the `context_stack` that is not building a
      graph function.
  (3) The gradient tape is paused while the scope is active.

In (2), a context is defined as either a graph or an eager context. Every
context switch, i.e., every installation of a graph as the default graph and
every switch into eager mode, is logged in a thread-local stack called
the `context_stack`; the log entry for a context switch is popped from the
stack when the context is exited. Entering an `init_scope`, with respect to
(2), is equivalent to crawling up the `context_stack`, finding the first
context that is not building a graph function, and entering it.

PiperOrigin-RevId: 179104270