tfdbg: Add DebugGradientRefIdentity op
Previously, the gradients debugger used DebugGradientIdentity regardless of
whether the input tensor is a reference type or non-reference type. This CL
switches it to a more specific DebugGradientRefIdentity type for reference
type tensors. This ensures that the output of the debug-identity op matches
its input exactly.

DebugGradientIdentity continues to be used for non-reference type tensors.

PiperOrigin-RevId: 181174976