Add zero padding option to `tf.image.extract_glimpse`
There has been several feature request in 16663 and 2134
asking for zero padding option to `tf.image.extract_glimpse`.
This fix is an attempt to address the issue through:
1. Add an additional args of `noise` so that it is possible
to specify `uniform`, `gaussian`, or `zero`.
2. In case `zero` is specified, extract_glimpse will fill
padding with zeros.

Note this PR keeps the original arg `uniform_noise` intact
(unless it conflict with `noise`), though it is possible
to deprecate the `uniform_noise` if desired.

This fix fixes 16663.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>