QueueOptions size var data types to size_t
QueueOptions' max_batch_size and max_enqueued_batches are positive quantities, and when compared, in the code, with unsigned member functions, a warning is raised. By changing the data type from int to size_t, not only the meaning of the member variables are more aligned with their intent, but also the comparisons are done between unsigned integers, thus fixing the warnings.