`PartitionedCallOp`: An op for executing multi-device functions.
A `PartitionedCallOp` allows for execution of functions across multiple devices
but within a single process. It proceeds by placing and partitioning the graph
underlying a given function body, instantiating for each partitioned subgraph a
function. The yielded function shards, which together are equivalent to the
original function, are then executed.

`PartitionedCallOp` is not part of the public TensorFlow API.

PiperOrigin-RevId: 194173114