[XLA:GPU] Set default kernel unrolling factor to 4
NVidia GPUs have 4-wide vector stores, so this is the natural unrolling factor
there. After recent fixes LLVM properly vectorizes the stores, giving a decent
performance win across the board and a huge performance win for F16 kernels.

At the moment the unrolling has no cost model and only applies to loop fusions.
All of the loop fusions I looked at were memory bound, unrolling them is a
clear win. When we add a cost model and/or auto tune later it has to be at
least as good as unrolling all the loop fusions.

PiperOrigin-RevId: 194249951