Add test case for empty tensor with clip ops.
This fix is based on 19337 and 19338. The issue was that
previously an empty tensor for tf.clip_by_value on GPU
triggers a crash. The issue should have been fixed by 19338 and
the recent master. It makes sense to adds the test case for this issue.

This fix adds the test case.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>