Install h5py on ppc64le required for keras_app...
Recently pip2/3 install keras_applications==1.0.2 was added to
install_pip_packages.sh. This pulls in the dependency h5py which fails
to compile from source on ppc64le. (On x86 h5py is pre-compiled so
it installs just fine). To get h5py to install we need to install
the libhdf5-dev deb package, and symlink the shared libraries to the names
h5py is expecting. Finally we need to include the path to the hdf5.h
header file.