Make the HloDomainRemover pass more configurable
Previously we had two different function to normalize instructions
within a domain where one of them was specified inside the metadata
while the other one is passed into the domain remover.

This change unifies them to use the externally passed in function for
both usecase to make it possible to rewrite both of them from the caller
of the domain remover (to add special logic).

PiperOrigin-RevId: 204715075