Add additional Python info into two error messages.
If the experimental option is enabled, this will insert Python format information into two error messages in placer.cc. The error messages are emitted when Placer fails to assign a device for an operation. The additional info will cause the Python layer to output where the node was defined in Python.

PiperOrigin-RevId: 204770562