Push tensors from client to workers.
At times, a server cannot open a reverse connection to the client. This is
required when using the _Send/_Recv ops and the client needs to send a tensor
to the server (tensors are pulled). Instead, this adds a way to push the
tensors directly from the client.

Currently, pushing tensors always happens in sync mode.

PiperOrigin-RevId: 205888825