Remove code that is no longer valid post change away from lowering to while loops.
When if op was lowered via while loops, the inputs and outputs matched. When
changed to the conditional this no longer holds. Resource variables are appended
at the end of the inputs & outputs, use the count of resource variables in the
input to identify the outputs due to resource variables.

This still requires that all the variables are last in the input to the op and that every variable read/written to is an output of the if.

PiperOrigin-RevId: 208097738