Simplify BUILD rule for MKL transpose op.
There is no reason for outside dependents to make a distinction between the
Eigen or MKL transpose operation, as the substitution is transparent. There is
also no need for transpose_op.cc itself to be compiled differently based on
whether MKL is in use or not. Therefore we remove external dependencies on
:mkl_transpose_op and make :transpose_op depend on it if needed (i.e., if
using MKL). This is consistent with how other transparent MKL operations (e.g.
matmul) are built.

PiperOrigin-RevId: 211874336