[XLA:GPU] Refactor some code for fusion output handling.
Move routine ConstructIrArrayForOutputs to class IrEmitter so that it can be
used in classes IrEmitterNested and IrEmitterUnnested.

Move the code that stores the address of each individual output of a
multiple output fusion to the tuple buffer of the fusion to an overload
version of routine llvm_ir::EmitTuple so that we can reduce code duplication.

PiperOrigin-RevId: 211884483