Fix ps0 OOM when workers too many.
When training with MonitoredTrainingSession in distributed mode,
all ops related with report_uninitialized_xxx will be placed on ps0,
which may lead to OOM when workers up to thousands.

We use a environment variable to solve it. When OOM happens, users can
set the local device to fix the problem, such as
os.environ['TF_LOCAL_DEVICE'] = '/job:worker/task:index'.