Support returning Operations from cond_v2.
This change mimics the behavior of the current tf.cond implementation,
which returns the predicate value for each Operation returned
(specifically, it returns the pivot value). We may want to revisit
this behavior for 2.0, e.g. returning None or disallowing it entirely,
but we'd like to continue supporting it in compat.v1.

PiperOrigin-RevId: 217930557