Make _XlaCompile resilient against "unimplemented" compiler failures
This CL makes _XlaCompile to fall back to the TF executor if the XLA compiler
can't compile the cluster for some reason.

While at it, make most of the fields in the XLA op kernels const so that we
don't accidentally mutate them in the ::Compute functions (this would result in
a race).

Also remove the `#undef OP_REQUIRES_OK_RETURN`; it has minimal value in a .cc
file.

PiperOrigin-RevId: 220156284