[XLA] Account for subcomputations' individual instructions in xla_hlo_profile.
Previously we'd say that the "cost" of an instruction with subcomputations
(e.g. a while loop) was equal to the total cost of its instructions.  Fine (I
mean, not actually fine, a loop can run multiple times, but at least it's
something), but we also profile *the inside* of a subcomputation, and we
weren't saving any information about *those* instructions.

PiperOrigin-RevId: 220381684