[XLA] Flatten tuples within while loops.
This is useful because it reduces the number of kTuple ops within the loop,
which can be expensive.  (Normally these would be removed by tuple simplifier,
but that doesn't operate across computation boundaries.)  But it's also useful
because it unlocks further optimization opportunities.  For example, it makes
it easier to remove dead loop parameters.

PiperOrigin-RevId: 220846222