Enable a deterministic and sequential execution order of potentially concurrent collective instances.
Before this change, collective op instances that were concurrent could execute
in a random, overlapping order across devices and workers.

This change introduces a mechanism to enforce the same execution order of
collective instances across devices and workers.  It first determines
dependencies between collective op nodes in the graph.  For each pair of
collective instances that do not depend on each other, this change adds an
arbitrary but deterministic control edge between the nodes.

This ordering only kicks in if `collective_deterministic_sequential_execution`
is true in ConfigProto.

PiperOrigin-RevId: 226397542