Remove unnecessary messing with colocation root device.
ColocateNodes chooses the which of the two colocation groups' roots
should become the new root. Then, it merges the device_names of the
two roots and assigns the merged name to the new root. The merged name
is always more (or equal) specific than the original names.

The removed code did the same thing with the only exception of
sometimes changing the device_name of the node that will not be the new
root. This has no effect because device selection for all the nodes in
a group is done using the information in the root of the group.

PiperOrigin-RevId: 231004097