[XLA:Python] Don't require a Backend to call ComputationBuilder.Build(), instead require it on Computation.Compile().
XLA computations can be built irrespective of what backends exist ? ultimately they are building a protocol buffer and we do not need a backend or XLA client of any kind to build one.

Leave code to forward a backend from Build() to Compile(); this can be removed after JAX is updated to pass the backend to Compile().

PiperOrigin-RevId: 233779679