[XLA] Always run DynamicUpdateSlice check in CanShareOperandBufferWithUser()
Originally, the passed-in function was supposed to only override the target-independent code to forbid sharing. But, at some point, this behavior was changed, and now the passed-in function completely overrides the target-independent code. We'd like to move back to the original formulation, but that would require a bit more surgery, since some of the constraints in the target-independent path are actually target-specific.

In the meanwhile, make one "obvious" check override the target if the check fails.

PiperOrigin-RevId: 236734154