Properly handle output of while loop in scoped allocator optimizer.
Before this change, if the output of a while loop, i.e. the Exit node, was an
input to a collective reduce, the scoped allocator optimizer would mark the
Exit node as allocated by scoped allocator.  This would cause the Exit node to
get inputs from different frames: the loop body has the loop's frame, but the
scoped allocator node is not in that frame.  This failed control flow
invariants.

After this change, scoped allocator optimizer inserts an identity node after
Exit and marks the identity for allocation via scoped allocator.  This works
around the problem because all scoped allocator related nodes are completely
outside the while frame.

PiperOrigin-RevId: 244093855