Add placer_inspection_required_ops_utils for dealing with ops calling functions
Operations calling functions are becoming ubiquitous in TF 2.0.
Examples include PartitionedCallOp, functional If/While, and Dataset ops.
We are calling such operations "deep ops".

When such ops are calling functions returning resources on multiple devices,
Placer need to look inside them. placer_inspection_required_ops_utils file
contains some utilities for placer to use including:

- PlacerInspectionRequiredOpChecker: A simple class with a single
  IsPlacerInspectionRequired method. It is a class instead of a standalone
  method because checking whether a function returns a resource takes some time
  and we cache the results.
- IsolatePlacerInspectionRequiredOps: This function adds Identity ops for each
  input/output of ops requiring placer inspection. It greatly simplifies
  the implementation of placing such ops.

The test file does not contain tests for IsolatePlacerInspectionRequiredOps()
because the following CLs add an IsolatePlacerInspectionRequiredOpsPass,
which is a thin wrapper on top of IsolatePlacerInspectionRequiredOps(), and
include a fair number of tests for it.

PiperOrigin-RevId: 244228697