Build efficient all-reduce ring for TPU model parallelism.
For non-model parallelism case, TF2XLA bridge reorders the replica ids to build an efficient all-reduce ring.
For model parallelism case, we disabled this reordering for correctness since we don't want to mess up user's input pipeline or weight shading, etc (see go/tpu-ids).

This guarantees the correctness but it sacrifices the performance, advanced model parallelism users(for example MeshTensorFlow) chose to build the efficient ring by themselves.

With this cl, we are going to move the ring build logic to device_assignment. This usually happens at the beginning of TF graph constructing, user will know how we map the replica id to physical coordinate and build the input-pipeline accordingly.

PiperOrigin-RevId: 244773805