[XLA] Fix the document on kConditional instruction.
The kConditional instruction has two forms: predicated kConditional and indexed
kConditional. The document says that the indexed kConditional form allows the
first operand to be a scalar of PRED, which is not what the hlo_parser and the
HLO backends support. Given the fact that predicated kConditional supports PRED
as the first operand, there is no need to allow the first operand of indexed
kConditional to be a scalar of PRED. Fix the document for these reasons.

Fix the XLA client API implementation for conditional to match the document.

Add an hlo_parser test case.

PiperOrigin-RevId: 246444546