Do not mark Min and Max as commutative, as they are not so w.r.t. NaN propagation.
This is because std::min(a, b) is equivalent to

template<class T>
const T& min(const T& a, const T& b)
{
    return (b < a) ? b : a;
}

So if either is NaN, the condition is always false and a is always returned; i.e. std::min(nan, 1) = nan, std::min(1, nan) = 1.

The Intel SIMD instructions for min and max, for example, have the same (non-commutative) behavior.

PiperOrigin-RevId: 246869272