[XLA] Make DISABLED_MANIFEST ignore shard numbers.
Right now you can disable a specific shard but not a sharded test.  This flips
it around; now you can disable a whole sharded test, but not a specific shard.

I think this is basically the right thing.  Even if you *wanted* to disable a
specific shard, sharding is unstable, so You Shouldn't Do That.

PiperOrigin-RevId: 247223725