[XLA:CPU] Prefix XLA:CPU custom-call macros with "XLA_CPU_".
In preparation for adding XLA:GPU custom-call support.

It's not possible to write a C++ function that implements a custom-call for
both CPU and GPU; for one thing, the GPU version needs to take a CUDA stream.
So it wouldn't make sense to use one macro for both CPU and GPU.

PiperOrigin-RevId: 247489592