[XLA] Be slightly more relaxed about denormals in exhaustive_op_test.
Given subnormal x previously the test would accept any of the following as a
"correct" value for f(x).

 - f(x) evaluated at full precision
 - f(copysign(0, x)), i.e. flush denormals to zero

Now we add a third option:

 - f(copysign(minfloat, x)), i.e. flush denormals to the closest *normal* value

This matches the behavior of XLA:CPU's log implementation when it's run with
ftz disabled on the CPU.  It also seems like a reasonable behavior, especially
for log, where this behavior avoids an infinity, which may avoid a nan down the
road.

PiperOrigin-RevId: 247766690